+++ js/Demo/layers/c.js
@@ -20,6 +20,7 @@ addLayer("c", {
         color: "#4BDC13",
         requires: new Decimal(10), // Can be a function that takes requirement increases into account
         resource: "lollipops", // Name of prestige currency
+        singular: "lollipop",
         baseResource: "points", // Name of resource prestige is based on
         baseAmount() {return player.points}, // Get the current amount of baseResource
         type: "normal", // normal: cost to gain currency depends on amount gained. static: cost depends on how much you already have